Tidy ComponentUtilities, Migrate ColumnSelectQueryString
