+++ src/Traits/Configuration/QueryStringConfiguration.php
@@ -4,13 +4,6 @@ namespace Rappasoft\LaravelLivewireTables\Traits\Configuration;
 
 trait QueryStringConfiguration
 {
-    public function setQueryStringAlias(string $queryStringAlias): self
-    {
-        $this->queryStringAlias = $queryStringAlias;
-
-        return $this;
-    }
-
     public function setupQueryStringStatus(): void
     {
         if (! $this->hasQueryStringStatus()) {
@@ -20,7 +13,7 @@ trait QueryStringConfiguration
             }
         }
     }
-
+    
     public function setQueryStringStatus(bool $status): self
     {
         $this->queryStringStatus = $status;
@@ -41,4 +34,30 @@ trait QueryStringConfiguration
 
         return $this;
     }
+
+    public function setQueryStringAlias(string $queryStringAlias): self
+    {
+        $this->queryStringAlias = $queryStringAlias;
+
+        return $this;
+    }
+
+    protected function setQueryStringConfig(string $type, array $config): self
+    {
+        $this->queryStringConfig[$type] = array_merge($this->getQueryStringConfig($type), $config);
+
+        return $this;
+    }
+
+    protected function setQueryStringConfigStatus(string $type, bool $status): self
+    {
+        return $this->setQueryStringConfig($type, ['status' => $status]);
+
+    }
+
+    protected function setQueryStringConfigAlias(string $type, string $alias): self
+    {
+        return $this->setQueryStringConfig($type, ['alias' => $alias]);
+    }
+
 }
