+++ src/Traits/Helpers/QueryStringHelpers.php
@@ -35,4 +35,30 @@ trait QueryStringHelpers
     {
         return $this->queryStringAlias ?? $this->getTableName();
     }
+
+    protected function getQueryStringConfig(string $type): array
+    {
+        return array_merge(['status' => null, 'alias' => null], ($this->queryStringConfig[$type] ?? []));
+    }
+
+    protected function hasQueryStringConfigStatus(string $type): bool
+    {
+        return isset($this->getQueryStringConfig($type)['status']);
+    }
+
+    protected function getQueryStringConfigStatus(string $type): bool
+    {
+        return $this->getQueryStringConfig($type)['status'] ?? $this->getQueryStringStatus();
+    }
+
+    protected function hasQueryStringConfigAlias(string $type): bool
+    {
+        return isset($this->getQueryStringConfig($type)['alias']);
+    }
+
+    protected function getQueryStringConfigAlias(string $type): string
+    {
+        return $this->getQueryStringConfig($type)['alias'] ?? $this->getQueryStringAlias().'-'.$type;
+    }
+
 }
