+++ tests/Unit/Traits/Core/QueryStrings/QueryStringForFiltersTest.php
@@ -6,107 +6,63 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Depends;
 use PHPUnit\Framework\Attributes\Group;
 use Rappasoft\LaravelLivewireTables\Tests\Http\Livewire\PetsTable;
-use Rappasoft\LaravelLivewireTables\Tests\TestCase;
 
 #[Group('Filters')]
-final class QueryStringForFiltersTest extends TestCase
+#[Group('QueryString')]
+final class QueryStringForFiltersTest extends QueryStringTestBase
 {
+    protected function setUp(): void
+    {
+        parent::setUp();
+    }
+
     public function test_can_get_default_filter_query_string_status(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
+        parent::$mock->configure();
+        parent::$mock->boot();
 
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-            }
-        };
-
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame(true, $mock->getQueryStringStatusForFilter());
+        $this->assertSame(true, parent::$mock->getQueryStringStatusForFilter());
     }
 
     public function test_can_disable_filter_query_string_status(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
-
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-                $this->setQueryStringForFilterDisabled();
-            }
-        };
-
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame(false, $mock->getQueryStringStatusForFilter());
+        parent::$mock->configure();
+        parent::$mock->boot();
+        $this->assertSame(true, parent::$mock->getQueryStringStatusForFilter());
+        parent::$mock->setQueryStringForFilterDisabled();
+        $this->assertSame(false, parent::$mock->getQueryStringStatusForFilter());
     }
 
     public function test_can_enable_filter_query_string_status(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
+        parent::$mock->configure();
+        parent::$mock->boot();
 
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-                $this->setQueryStringForFilterDisabled();
-            }
-        };
-
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame(false, $mock->getQueryStringStatusForFilter());
-        $mock->setQueryStringForFilterEnabled();
-        $this->assertSame(true, $mock->getQueryStringStatusForFilter());
+        $this->assertSame(true, parent::$mock->getQueryStringStatusForFilter());
+        parent::$mock->setQueryStringForFilterDisabled();
+        $this->assertSame(false, parent::$mock->getQueryStringStatusForFilter());
+        parent::$mock->setQueryStringForFilterEnabled();
+        $this->assertSame(true, parent::$mock->getQueryStringStatusForFilter());
 
     }
 
     public function test_can_get_default_filter_query_string_alias(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
+        parent::$mock->configure();
+        parent::$mock->boot();
 
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-            }
-        };
-
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame('table-filters', $mock->getQueryStringAliasForFilter());
+        $this->assertSame('table-filters', parent::$mock->getQueryStringAliasForFilter());
 
     }
 
     public function test_can_change_default_filter_query_string_alias(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
+        parent::$mock->configure();
+        parent::$mock->boot();
 
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-            }
-        };
-
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame('table-filters', $mock->getQueryStringAliasForFilter());
-        $mock->setQueryStringAliasForFilter('pet-filters');
-        $this->assertSame('pet-filters', $mock->getQueryStringAliasForFilter());
-        $this->assertTrue($mock->hasQueryStringAliasForFilter());
+        $this->assertSame('table-filters', parent::$mock->getQueryStringAliasForFilter());
+        parent::$mock->setQueryStringAliasForFilter('pet-filters');
+        $this->assertSame('pet-filters', parent::$mock->getQueryStringAliasForFilter());
+        $this->assertTrue(parent::$mock->hasQueryStringAliasForFilter());
     }
 }
