+++ tests/Unit/Traits/Core/QueryStrings/QueryStringForSortTest.php
@@ -4,107 +4,68 @@ namespace Rappasoft\LaravelLivewireTables\Tests\Unit\Traits\Core\QueryStrings;
 
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Depends;
+use PHPUnit\Framework\Attributes\Group;
 use Rappasoft\LaravelLivewireTables\Tests\Http\Livewire\PetsTable;
-use Rappasoft\LaravelLivewireTables\Tests\TestCase;
 
-final class QueryStringForSortTest extends TestCase
+#[Group('QueryString')]
+final class QueryStringForSortTest extends QueryStringTestBase
 {
+    protected function setUp(): void
+    {
+        parent::setUp();
+    }
+
     public function test_can_get_default_sort_query_string_status(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
+        parent::$mock->configure();
+        parent::$mock->boot();
 
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-            }
-        };
-
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame(true, $mock->getQueryStringStatusForSort());
+        $this->assertSame(true, parent::$mock->getQueryStringStatusForSort());
     }
 
     public function test_can_disable_sort_query_string_status(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
+        parent::$mock->configure();
+        parent::$mock->boot();
+        parent::$mock->setQueryStringForSortDisabled();
 
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-                $this->setQueryStringForSortDisabled();
-            }
-        };
 
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame(false, $mock->getQueryStringStatusForSort());
+        $this->assertSame(false, parent::$mock->getQueryStringStatusForSort());
     }
 
     public function test_can_enable_sort_query_string_status(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
+        parent::$mock->configure();
+        parent::$mock->boot();
+        parent::$mock->setQueryStringForSortDisabled();
 
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-                $this->setQueryStringForSortDisabled();
-            }
-        };
 
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame(false, $mock->getQueryStringStatusForSort());
-        $mock->setQueryStringForSortEnabled();
-        $this->assertSame(true, $mock->getQueryStringStatusForSort());
+        $this->assertSame(false, parent::$mock->getQueryStringStatusForSort());
+        parent::$mock->setQueryStringForSortEnabled();
+        $this->assertSame(true, parent::$mock->getQueryStringStatusForSort());
 
     }
 
     public function test_can_get_default_sort_query_string_alias(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
+        parent::$mock->configure();
+        parent::$mock->boot();
+        parent::$mock->setQueryStringForSortDisabled();
 
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-            }
-        };
-
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame('table-sorts', $mock->getQueryStringAliasForSort());
+        $this->assertSame('table-sorts', parent::$mock->getQueryStringAliasForSort());
 
     }
 
     public function test_can_change_default_sort_query_string_alias(): void
     {
-        $mock = new class extends PetsTable
-        {
-            public ?array $testAttributesArray;
+        parent::$mock->configure();
+        parent::$mock->boot();
+        parent::$mock->setQueryStringForSortDisabled();
 
-            public function configure(): void
-            {
-                $this->setDataTableFingerprint('test');
-            }
-        };
 
-        $mock->configure();
-        $mock->boot();
-
-        $this->assertSame('table-sorts', $mock->getQueryStringAliasForSort());
-        $mock->setQueryStringAliasForSort('pet-sorts');
-        $this->assertSame('pet-sorts', $mock->getQueryStringAliasForSort());
-        $this->assertTrue($mock->hasQueryStringAliasForSort());
+        $this->assertSame('table-sorts', parent::$mock->getQueryStringAliasForSort());
+        parent::$mock->setQueryStringAliasForSort('pet-sorts');
+        $this->assertSame('pet-sorts', parent::$mock->getQueryStringAliasForSort());
+        $this->assertTrue(parent::$mock->hasQueryStringAliasForSort());
     }
 }
